+++ src/rules/index.ts
@@ -1,1 +1,3 @@
 export { noInlineInterfaceObjectTypesRule, RULE_NAME as noInlineInterfaceObjectTypesRuleName } from './no-inline-interface-object-types';
+
+export { preferLoggerRule, RULE_NAME as preferLoggerRuleName } from './prefer-logger';